requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain
langchain-google-genai
chromadb
tiktoken
openai
pydantic
google-generativeai
